requirements.txt
Flask==3.1.0
numpy==2.1.3
Pillow==11.0.0
tensorflow==2.19.0
